[1222] Implement AP search in support
