+++ app/components/search_result_title_component.html.erb
