+++ app/components/tab_navigation.html.erb
@@ -1,4 +1,4 @@
-<nav class="app-tab-navigation govuk-!-margin-bottom-4" aria-label="Sub navigation">
+<nav class="app-tab-navigation govuk-!-margin-bottom-7" aria-label="Sub navigation">
   <ul class="app-tab-navigation__list">
     <% items.each do |item| %>
       <li class="app-tab-navigation__item">
