+++ app/forms/accredited_provider_search_form.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class AccreditedProviderSearchForm
+  include ActiveModel::Model
+
+  FIELDS = %i[
+    query
+  ].freeze
+
+  attr_accessor(*FIELDS)
+
+  validates :query, presence: true, length: { minimum: 2 }
+end
