+++ app/views/support/providers/accredited_provider_search/new.html.erb
@@ -1,0 +1,32 @@
+
+<% content_for :page_title, title_with_error_prefix(t(".title"), @accredited_provider_search_form.errors.any?) %>
+
+<% content_for :before_content do %>
+  <%= govuk_back_link_to(support_recruitment_cycle_provider_accredited_providers_path) %>
+<% end %>
+
+<div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+
+    <%= form_with(
+        model: @accredited_provider_search_form,
+        url: search_support_recruitment_cycle_provider_accredited_providers_path,
+        method: :post
+      ) do |f| %>
+
+      <%= f.govuk_error_summary %>
+
+      <%= f.govuk_text_field(
+        :query,
+        label: { text: t(".title"), size: "l", tag: "h1" },
+        caption: { text: t(".caption", provider_name: @provider.provider_name, code: @provider.provider_code), size: "l" }
+      ) %>
+
+      <%= f.govuk_submit t("continue") %>
+    <% end %>
+
+    <p class="govuk-body">
+      <%= govuk_link_to(t("cancel"), support_recruitment_cycle_provider_accredited_providers_path) %>
+    </p>
+  </div>
+</div>
