+++ app/views/support/providers/accredited_providers/index.html.erb
@@ -1,0 +1,3 @@
+<%= render PageTitle.new(title: "support.providers.accredited_providers.index") %>
+
+<%= govuk_link_to("Add accredited provider", search_support_recruitment_cycle_provider_accredited_providers_path, class: "govuk-button govuk-!-margin-bottom-3") %>
