+++ config/locales/en.yml
@@ -166,6 +166,8 @@ en:
             check: "Check your answers"
             show: "User overview"
             delete: "Are you sure you want to remove this user?"
+          accredited_providers:
+            index: Accredited providers
           onboardings:
             new: "Organisation details"
         users:
@@ -270,10 +272,9 @@ en:
           title: Enter a school, university, college, URN or postcode
           caption: *add_school
           cannot_find: I cannot find the school - enter manually
-    schools:
-      search_result_title_component:
-        add_school: *add_school
-        many_results_html: Showing the first 15 results. %{link} if the school you're looking for is not listed.
+  search_result_title_component:
+    add_school: *add_school
+    many_results_html: Showing the first 15 results. %{link} if the school you're looking for is not listed.
   support:
     data_exports:
       index:
@@ -307,6 +308,10 @@ en:
         new: "Add schools"
         show: "Add school (%{position} of %{max})"
         check: "Check your answers"
+      accredited_provider_search:
+        new:
+          title: Enter a provider name, UKPRN or postcode
+          caption: "Add accredited provider - %{provider_name} (%{code})"
     flash:
       created: "%{resource} successfully created"
       updated: "%{resource} successfully updated"
@@ -469,6 +474,15 @@ en:
   activemodel:
     errors:
       models:
+        accredited_provider_search_form:
+          attributes:
+            query:
+              blank: Enter a provider name, UKPRN or postcode
+              too_short: Enter at least two characters
+        accredited_provider_select_form:
+          attributes:
+            provider_id:
+              blank: Select an accredited provider
         publish/engineers_teach_physics_form:
           attributes:
             campaign_name:
